+++ packages/decap-cms-core/src/lib/__tests__/formatters.spec.js
@@ -247,6 +247,27 @@ describe('formatters', () => {
         'Ignoring unknown variable “author-email” in open authoring message template.',
       );
     });
+
+    it('should return commit with trailer when signoff_commits is enabled', () => {
+      const collection = Map({ label_singular: 'Collection' });
+      const config = {
+        backend: {
+          signoff_commits: true,
+        },
+      };
+
+      expect(
+        commitMessageFormatter('create', config, {
+          slug: 'doc-slug',
+          path: 'file-path',
+          collection,
+          authorName: 'Test User',
+          authorEmail: 'test-user@example.org',
+        }),
+      ).toEqual(
+        'Create Collection “doc-slug”\n\nSigned-off-by: Test User <test-user@example.org>\n',
+      );
+    });
   });
 
   describe('prepareSlug', () => {
@@ -274,8 +295,8 @@ describe('formatters', () => {
   };
 
   describe('slugFormatter', () => {
-    const date = new Date('2020-01-01').valueOf();
-    Date.now = jest.spyOn(Date, 'now').mockImplementation(() => date);
+    const date = new Date('2020-01-01T13:28:27.679Z').valueOf();
+    jest.spyOn(Date, 'now').mockImplementation(() => date);
 
     const { selectIdentifier } = require('../../reducers/collections');
 
@@ -312,10 +333,22 @@ describe('formatters', () => {
       ).toBe('entry-slug');
     });
 
+    it('should allow filters in slug templates', () => {
+      selectIdentifier.mockReturnValueOnce('published');
+
+      expect(
+        slugFormatter(
+          Map({ slug: "{{published | date('MM-DD')}}" }),
+          Map({ title: 'Post Title', published: new Date(date) }),
+          slugConfig,
+        ),
+      ).toBe('01-01');
+    });
+
     it('should see date filters applied to date from entry if it exists', () => {
       const { selectInferredField } = require('../../reducers/collections');
       selectInferredField.mockReturnValue('date');
-      const entryDate = new Date('2026-10-20');
+      const entryDate = new Date('2026-10-20T13:28:27.679Z');
 
       expect(
         slugFormatter(
@@ -329,7 +362,7 @@ describe('formatters', () => {
     it('should see date filters applied to publishDate from entry if it exists', () => {
       const { selectInferredField } = require('../../reducers/collections');
       selectInferredField.mockReturnValue('publishDate');
-      const entryDate = new Date('2026-10-20');
+      const entryDate = new Date('2026-10-20T13:28:27.679Z');
 
       expect(
         slugFormatter(
@@ -565,6 +598,66 @@ describe('formatters', () => {
       expect(console.error).toHaveBeenCalledWith(
         'Collection "posts" configuration error:\n  `preview_path_date_field` must be a field with a valid date. Ignoring `preview_path`.',
       );
+    });
+
+    it('should preserve slashes in value when configured', () => {
+      expect(
+        previewUrlFormatter(
+          'https://www.example.com',
+          Map({
+            preview_path: 'prefix/{{value}}',
+            preview_path_preserve_slashes: true,
+          }),
+          'backendSlug',
+          Map({ data: Map({ value: 'nested/value' }) }),
+          slugConfig,
+        ),
+      ).toBe('https://www.example.com/prefix/nested/value');
+    });
+
+    it('should sanitize slashes in value when not configured', () => {
+      expect(
+        previewUrlFormatter(
+          'https://www.example.com',
+          Map({
+            preview_path: 'prefix/{{value}}',
+          }),
+          'backendSlug',
+          Map({ data: Map({ value: 'nested/value' }) }),
+          slugConfig,
+        ),
+      ).toBe('https://www.example.com/prefix/nested-value');
+    });
+
+    it('should preserve slashes in value for nested collections by default', () => {
+      expect(
+        previewUrlFormatter(
+          'https://www.example.com',
+          Map({
+            preview_path: 'prefix/{{value}}',
+            nested: { depth: 10 },
+          }),
+          'backendSlug',
+          Map({ data: Map({ value: 'nested/value' }) }),
+          slugConfig,
+        ),
+      ).toBe('https://www.example.com/prefix/nested/value');
+    });
+
+    it('should sanitize slashes in value for nested collections when explicitly disabled', () => {
+      expect(
+        previewUrlFormatter(
+          'https://www.example.com',
+          Map({
+            preview_path: 'prefix/{{value}}',
+            nested: { depth: 10 },
+            preview_path_preserve_slashes: false,
+          }),
+          'backendSlug',
+          Map({ data: Map({ value: 'nested/value' }) }),
+          slugConfig,
+        ),
+      ).toBe('https://www.example.com/prefix/nested-value');
     });
   });