+++ packages/decap-cms-lib-auth/src/utils.js
@@ -1,7 +1,5 @@
-import { v4 as uuid } from 'uuid';
-
 export function createNonce() {
-  const nonce = uuid();
+  const nonce = crypto.randomUUID();
   window.sessionStorage.setItem('decap-cms-auth', JSON.stringify({ nonce }));
   return nonce;
 }
@@ -9,7 +7,7 @@ export function createNonce() {
 export function validateNonce(check) {
   const auth = window.sessionStorage.getItem('decap-cms-auth');
   const valid = auth && JSON.parse(auth).nonce;
-  window.localStorage.removeItem('decap-cms-auth');
+  window.sessionStorage.removeItem('decap-cms-auth');
   return check === valid;
 }