+++ packages/decap-cms-locales/src/sl/index.js
@@ -42,7 +42,8 @@ const sl = {
     collectionTop: {
       sortBy: 'Razvrsti po',
       viewAs: 'Poglej kot',
-      newButton: 'Nov %{collectionLabel}',
+      newButton: '＋ %{collectionLabel}',
+      newButtonAriaLabel: 'Ustvari vnos vrste %{collectionLabel}',
       ascending: 'Naraščajoče',
       descending: 'Padajoče',
       searchResults: 'Rezultati iskanja za "%{searchTerm}"',