+++ packages/decap-cms-widget-list/src/ListControl.js
@@ -7,7 +7,6 @@ import { List, Map, fromJS } from 'immutable';
 import partial from 'lodash/partial';
 import isEmpty from 'lodash/isEmpty';
 import uniqueId from 'lodash/uniqueId';
-import { v4 as uuid } from 'uuid';
 import DecapCmsWidgetObject from 'decap-cms-widget-object';
 import {
   DndContext,
@@ -211,7 +210,7 @@ export default class ListControl extends React.Component {
     const { field, value } = props;
     const listCollapsed = field.get('collapsed', true);
     const itemsCollapsed = (value && Array(value.size).fill(listCollapsed)) || [];
-    const keys = (value && Array.from({ length: value.size }, () => uuid())) || [];
+    const keys = (value && Array.from({ length: value.size }, () => crypto.randomUUID())) || [];
 
     this.state = {
       listCollapsed,
@@ -353,7 +352,7 @@ export default class ListControl extends React.Component {
     const { value, onChange, field } = this.props;
     const addToTop = field.get('add_to_top', false);
 
-    const itemKey = uuid();
+    const itemKey = crypto.randomUUID();
     this.setState({
       itemsCollapsed: addToTop
         ? [false, ...this.state.itemsCollapsed]