+++ packages/decap-cms-widget-richtext/src/RichtextControl/VisualEditor.js
@@ -28,11 +28,13 @@ import LinkElement from './components/Element/LinkElement';
 import ImageElement from './components/Element/ImageElement';
 import ExtendedBlockquotePlugin from './plugins/ExtendedBlockquotePlugin';
 import ImagePlugin from './plugins/ImagePlugin';
+import BreakPlugin from './plugins/BreakPlugin';
 import ShortcodePlugin from './plugins/ShortcodePlugin';
 import { TablePlugin, TableRowPlugin, TableCellPlugin } from './plugins/TablePlugin';
 import defaultEmptyBlock from './defaultEmptyBlock';
 import { mergeMediaConfig } from './mergeMediaConfig';
 import { handleLinkClick } from './linkHandler';
+import { handlePasteHtml } from './pasteHandler';
 
 function editorStyles({ minimal }) {
   return css`
@@ -87,6 +89,10 @@ export default function VisualEditor(props) {
     onChange(mdValue);
   }
 
+  function handlePaste(event) {
+    handlePasteHtml({ event, editor, isDisabled });
+  }
+
   const initialValue = props.value
     ? markdownToSlate(props.value, { editorComponents, voidCodeBlock: !!codeBlockComponent })
     : emptyValue;
@@ -133,6 +139,7 @@ export default function VisualEditor(props) {
         shortcuts: { toggle: { keys: 'mod+shift+c' } },
       }),
       ListPlugin,
+      BreakPlugin,
       ImagePlugin,
       LinkPlugin.configure({
         node: { component: LinkElement },
@@ -188,7 +195,7 @@ export default function VisualEditor(props) {
               />
             </EditorControlBar>
             <div css={editorStyles({ minimal: field.get('minimal') })}>
-              <Editor isDisabled={isDisabled} />
+              <Editor isDisabled={isDisabled} onPaste={handlePaste} />
             </div>
           </Plate>
         </div>