+++ packages/decap-cms-lib-auth/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.2.0](https://github.com/decaporg/decap-cms/compare/decap-cms-lib-auth@3.1.0...decap-cms-lib-auth@3.2.0) (2026-06-08)
+
+**Note:** Version bump only for package decap-cms-lib-auth
+
+# [3.1.0](https://github.com/decaporg/decap-cms/compare/decap-cms-lib-auth@3.0.6...decap-cms-lib-auth@3.1.0) (2026-06-01)
+
+**Note:** Version bump only for package decap-cms-lib-auth
+
 ## [3.0.6](https://github.com/decaporg/decap-cms/compare/decap-cms-lib-auth@3.0.5...decap-cms-lib-auth@3.0.6) (2025-07-31)
 
 **Note:** Version bump only for package decap-cms-lib-auth