+++ packages/decap-cms/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.14.0](https://github.com/decaporg/decap-cms/compare/decap-cms@3.13.0...decap-cms@3.14.0) (2026-06-08)
+
+**Note:** Version bump only for package decap-cms
+
+# [3.13.0](https://github.com/decaporg/decap-cms/compare/decap-cms@3.12.2...decap-cms@3.13.0) (2026-06-01)
+
+**Note:** Version bump only for package decap-cms
+
+## [3.12.2](https://github.com/decaporg/decap-cms/compare/decap-cms@3.12.1...decap-cms@3.12.2) (2026-04-17)
+
+**Note:** Version bump only for package decap-cms
+
+## [3.12.1](https://github.com/decaporg/decap-cms/compare/decap-cms@3.12.0...decap-cms@3.12.1) (2026-04-16)
+
+**Note:** Version bump only for package decap-cms
+
+# [3.12.0](https://github.com/decaporg/decap-cms/compare/decap-cms@3.11.0...decap-cms@3.12.0) (2026-04-16)
+
+**Note:** Version bump only for package decap-cms
+
 # [3.11.0](https://github.com/decaporg/decap-cms/compare/decap-cms@3.10.1...decap-cms@3.11.0) (2026-03-24)
 
 ### Features